pyproject.toml
[project]
name = "scheduler-wat-v2-2024"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "colorama>=0.4.6",
    "flask>=3.1.3",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "gunicorn>=25.3.0",
    "protobuf>=7.34.1",
    "pytz>=2026.1.post1",
    "requests>=2.33.1",
    "rich>=14.3.3",
    "werkzeug>=3.1.8",
    "flasgger>=0.9.7.1",
]
